pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openresponses-impl-client-google"
version = "0.1.0"
description = "Python Client library implementing the OpenResponses specification"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
  { name = "rsato", email = "satookok@gmail.com" }
]
keywords = [
  "openresponses",
  "gemini",
  "google-ai",
  "llm",
  "ai-client",
  "genai",
  "responses-api"
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "google-genai>=1.72.0",
    "openresponses-impl-core>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/post-class/openresponses-impl-client-google"
Repository = "https://github.com/post-class/openresponses-impl-client-google"



[tool.setuptools.packages.find]
where = ["src"]
include = ["openresponses_impl_client_google*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["test"]

[dependency-groups]
dev = [
    "build>=1.4.3",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "twine>=6.2.0",
]
